+++ summoner-cli/CHANGELOG.md
@@ -3,7 +3,7 @@
 `summoner` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
-## Unreleased: 1.2.0
+## 1.2.0 — Nov 30, 2018
 
 * Make `cabal-version: 2.0` default in generated projects.
 * [#11](https://github.com/kowainik/summoner/issues/11):